requirements/common.txt
Django==2.1.3
pytz==2018.7
django-health-check==3.8.0
whitenoise==4.1.2
psycopg2==2.7.6.1
djangorestframework==3.9.1
celery[redis]==4.2.1
beautifulsoup4==4.6.3
requests==2.20.1
colorlog==4.0.2
django-webpack-loader==0.6.0
Pillow==5.4.1
